Allow each key to contain multiple values
There is no way to use a variable number of keys without reusing items
with different search key values, because every used key must be defined
both in the options and the data.

Allow the use of Arrays and Objects for keys, so that each item may have
distinct alternatives for its keys.