Improvements to TS typings (#253)
* Make the Fuse constructor accept ReadonlyArray

Fuse will not modify the input array, so a readonly array is ok as an
input. The `ReadonlyArray<any>` will also accept any `Array<any>`, so
this is not a breaking change.

* Make Fuse generic

Since the element type of the input array and the search result type are
always the same, it makes sense to make Fuse generic over that type.