Fix typings
This fixes the typings for nested keys which was previously broken, and
reported by @mmajko on #261

While this does fix the typings at least for now, they can be made even
more ideal. To achieve this however advanced typings would be required
and a request has been made to the utility-types repository (see:
https://github.com/piotrwitek/utility-types/issues/85) for the required
advanced type. To summarize that issue, if we would have a Flatten<T>
that would flatten the entire given typing T to single keys it would be
a perfect match for this library.